pyproject.toml
[project]
name = "ai-search-staatsarchiv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "numpy>=2.3.5",
    "pandarallel>=1.6.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "spacy>=3.8.11",
    "streamlit>=1.52.1",
    "tqdm>=4.67.1",
    "weaviate-client>=4.18.3",
    "de-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl",
    "lxml>=6.0.2",
]
